put urlToMedia into helperFunction file
