+++ src/components/Studio/ContentTypes/EmbedContentEdit.tsx
@@ -12,11 +12,12 @@ import { slideEmbedContentSchema } from '@/src/lib/validations/slidecontent'
 import { Input, InputLabel } from '@/src/components/Elements/Input'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useEffect, useState } from 'react'
-import { media_type, media_types } from '@/src/lib/media/media'
+import { media_type } from '@/src/lib/media/media'
 import { useTranslation } from '@/src/app/i18n/client'
 import { fallbackLng, languages } from '@/src/app/i18n/settings'
 import { Embed } from '../../embeds/Embed'
 import { SlideContent } from '@prisma/client'
+import { urlToMedia } from '../../HelperFunctions'
 
 interface EmbedContentEditProps extends React.HTMLAttributes<HTMLFormElement> {
   storyStepId: string
@@ -139,35 +140,4 @@ export function EmbedContentEdit({
       </div>
     </form>
   )
-}
-
-export function urlToMedia(url: string): media_type {
-  // check if url is a valid url with zod
-  try {
-    z.string().url().parse(url)
-
-    // check if url ir a known media url
-    for (var i = 0; i < media_types.length; i++) {
-      if (url.toString().match(media_types[i].match_str)) {
-        const media = { ...media_types[i] }
-        media.url = url
-        return media
-      }
-    }
-    // return unknown media
-    return {
-      type: 'unknown',
-      name: 'Unknown',
-      match_str: / /,
-      url: '',
-    }
-  } catch {
-    // return unknown media
-    return {
-      type: 'unknown',
-      name: 'Unknown',
-      match_str: / /,
-      url: '',
-    }
-  }
 }
